requirements.txt
# FastAPI backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# PDF processing
pypdf==3.17.1
pdfplumber==0.11.4
PyMuPDF==1.23.7

# RAG + embeddings
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# LLM providers
google-generativeai==0.5.2
openai==1.54.1
huggingface-hub==0.25.2

# LangChain stack
langchain==0.2.6
langchain-core==0.2.10
langchain-community==0.2.6
langchain-openai==0.1.7
langchain-google-genai==1.0.5

# Frontend
streamlit==1.38.0
requests==2.31.0

